Three files for the UMAP-kings under the sky.
