+++ src/component/index.ts
@@ -5,3 +5,4 @@ export * from './ContactMap';
 export * from './FeatureViewer';
 export * from './NGLComponent';
 export * from './TensorTComponent';
+export * from './UMAPVisualization';
