--- src/singlepage/index.ts
@@ -1,1 +1,0 @@
-export * from './UMAPVisualization';
